build: move utils directory
