+++ src/components/image-card/image-card.vue
@@ -70,15 +70,11 @@ import { computed, nextTick, ref } from 'vue'
 import type { ElInput } from 'element-plus'
 import { useRoute } from 'vue-router'
 import { useStore } from '@/store'
-import axios from '@/common/utils/axios'
+import axios from '@/utils/axios'
 import { UploadedImageModel } from '@/common/model/upload.model'
-import { getBase64ByImageUrl, getImage } from '@/common/utils/rename-image'
-import { uploadImage_single } from '@/common/utils/upload-helper'
-import {
-  getFilename,
-  getFileSize,
-  getFileSuffix
-} from '@/common/utils/file-handle-helper'
+import { getBase64ByImageUrl, getImage } from '@/utils/rename-image'
+import { uploadImage_single } from '@/utils/upload-helper'
+import { getFilename, getFileSize, getFileSuffix } from '@/utils/file-handle-helper'
 import ExternalLinkType from '@/common/model/external-link.model'
 
 const props = defineProps({
