+++ src/store/modules/user-config-info/index.ts
@@ -4,11 +4,11 @@ import {
   UserConfigInfoModel
 } from '@/common/model/user-config-info.model'
 import { PICX_CONFIG } from '@/common/model/storage.model'
-import { deepAssignObject, cleanObject } from '@/common/utils/object-helper'
+import { deepAssignObject, cleanObject } from '@/utils/object-helper'
 import UserConfigInfoStateTypes from '@/store/modules/user-config-info/types'
 import RootStateTypes from '@/store/types'
 import { DirModeEnum } from '@/common/model/dir.model'
-import TimeHelper from '@/common/utils/time-helper'
+import TimeHelper from '@/utils/time-helper'
 
 const initUserConfigInfo = (): UserConfigInfoModel => {
   const initConfig: UserConfigInfoModel = {
