+++ src/utils/compress.ts
