+++ src/utils/set-theme-mode.ts
